refactor: :recycle: move walkthrough overlay on event matching to different file
